Move UTs to a separate file
