+++ analyzers/tests/SonarAnalyzer.UnitTest/SymbolicExecution/Roslyn/RoslynSymbolicExecutionTest.Branching.cs
@@ -70,32 +70,6 @@ Tag(""End"");";
                 "AfterFinally");    // Dequeue after "if" branch
         }
 
-//         [TestMethod]
-//         public void Branching_Test()
-//         {
-//             const string code = @"
-// Tag(""Entry"");
-// var x = 3;
-// if (x != 1)
-// {
-//     Tag(""1"");
-//     if (x != 2)
-//     {
-//         Tag(""2"");
-//     }
-//     else
-//     {
-//         Tag(""-2"");
-//     }
-// }
-// else
-// {
-//     Tag(""-1"");
-// }
-// Tag(""End"");";
-//             SETestContext.CreateCS(code).Validator.ValidateTagOrder("Entry", "1", "-1", "End", "2", "-2");
-//         }
-
         [TestMethod]
         public void Branching_BlockProcessingOrder_VB()
         {
@@ -321,139 +295,6 @@ else
             validator.ValidateTag("ToString", x => x.HasConstraint(TestConstraint.First).Should().BeTrue());
             validator.ValidateTag("GetHashCode", x => x.HasConstraint(TestConstraint.First).Should().BeFalse()); // Nobody set the constraint on that path
             validator.ValidateExitReachCount(2);    // Once for each state
-        }
-
-        [TestMethod]
-        public void Finally_Simple()
-        {
-            const string code = @"
-Tag(""BeforeTry"");
-try
-{
-    Tag(""InTry"");
-}
-finally
-{
-    Tag(""InFinally"");
-}
-Tag(""AfterFinally"");
-";
-            SETestContext.CreateCS(code).Validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InFinally",
-                "AfterFinally");
-        }
-
-        [TestMethod]
-        public void Finally_Nested()
-        {
-            const string code = @"
-Tag(""BeforeTry"");
-try
-{
-    Tag(""InOuterTry"");
-    try
-    {
-        Tag(""InNestedTry"");
-    }
-    finally
-    {
-        Tag(""InNestedFinally"");
-    }
-}
-finally
-{
-    Tag(""InFinally"");
-}
-Tag(""AfterFinally"");
-";
-            SETestContext.CreateCS(code).Validator.ValidateTagOrder(
-                "BeforeTry",
-                "InOuterTry",
-                "InNestedTry",
-                "InNestedFinally",
-                "InFinally",
-                "AfterFinally");
-        }
-
-        [TestMethod]
-        public void Finally_Nested_InstructionAfterFinally()
-        {
-            const string code = @"
-Tag(""BeforeTry"");
-try
-{
-    Tag(""InFirstTry"");
-    try
-    {
-        Tag(""InNestedTry"");
-    }
-    finally
-    {
-        true.ToString();
-        Tag(""InNestedFinally"");
-    }
-    Tag(""AfterNestedFinally"");
-}
-finally
-{
-    Tag(""InFinally"");
-}
-Tag(""AfterFinally"");
-";
-            SETestContext.CreateCS(code).Validator.ValidateTagOrder(
-                "BeforeTry",
-                "InFirstTry",
-                "InNestedTry",
-                "InNestedFinally",
-                "AfterNestedFinally",
-                "InFinally",
-                "AfterFinally");
-        }
-
-        [TestMethod]
-        public void Finally_BranchInNested()
-        {
-            const string code = @"
-Tag(""BeforeTry"");
-try
-{
-    Tag(""InFirstTry"");
-    try
-    {
-        Tag(""InNestedTry"");
-        if (1 == 2)
-        {
-            Tag(""1"");
-        }
-        else
-        {
-            Tag(""2"");
-        }
-    }
-    finally
-    {
-        Tag(""InNestedFinally"");
-    }
-    Tag(""AfterNestedFinally"");
-}
-finally
-{
-    Tag(""InFinally"");
-}
-Tag(""AfterFinally"");
-";
-            SETestContext.CreateCS(code).Validator.ValidateTagOrder(
-                "BeforeTry",
-                "InFirstTry",
-                "InNestedTry",
-                "InNestedFinally",
-                "1",
-                "2",
-                "InFinally",
-                "AfterNestedFinally",
-                "AfterFinally");
         }
     }
 }
